refactor: break down kitCrudOperations.ts to meet 300 LOC standard
